+++ py/desisim/simexp.py
@@ -769,7 +769,7 @@ def get_mock_spectra(fiberassign, mockdir=None, nside=64):
         #- Sky fibers aren't in the truth files
         ok = ~np.in1d(targetids, skyids)
 
-        tmpflux, tmpwave, tmpmeta = read_mock_spectra(truthfile, targetids[ok])
+        tmpflux, tmpwave, tmpmeta, tmpobjmeta = read_mock_spectra(truthfile, targetids[ok])
 
         if flux is None:
             nwave = tmpflux.shape[1]
@@ -797,7 +797,7 @@ def read_mock_spectra(truthfile, targetids, mockdir=None):
     Reads mock spectra from a truth file
 
     Args:
-        truthfile (str): full path to a mocks spectra_truth*.fits file
+        truthfile (str): full path to a mocks truth-\*.fits file
         targetids (array-like): targetids to load from that file
         mockdir: ???
 
@@ -818,10 +818,22 @@ def read_mock_spectra(truthfile, targetids, mockdir=None):
     #     truth = fx['TRUTH'].data
     #     wave = fx['WAVE'].data
     #     flux = fx['FLUX'].data
+    objtruth = dict()
     with fitsio.FITS(truthfile) as fx:
         truth = fx['TRUTH'].read()
         wave = fx['WAVE'].read()
         flux = fx['FLUX'].read()
+
+        if 'OBJTYPE' in truth.dtype.names:
+            # output of desisim.obs.new_exposure
+            objtype = [oo.decode('ascii').strip().upper() for oo in truth['OBJTYPE']] 
+        else:
+            # output of desitarget.mock.build.write_targets_truth
+            objtype = [oo.decode('ascii').strip().upper() for oo in truth['TEMPLATETYPE']] 
+        for obj in set(objtype):
+            extname = 'TRUTH_{}'.format(obj)
+            if extname in fx:
+                objtruth[obj] = fx[extname].read()
 
     missing = np.in1d(targetids, truth['TARGETID'], invert=True)
     if np.any(missing):
@@ -832,10 +844,15 @@ def read_mock_spectra(truthfile, targetids, mockdir=None):
     ii = np.in1d(truth['TARGETID'], targetids)
     flux = flux[ii]
     truth = truth[ii]
+    if bool(objtruth):
+        for obj in objtruth.keys():
+            ii = np.in1d(objtruth[obj]['TARGETID'], targetids)
+            objtruth[obj][:] = objtruth[obj][ii]
 
     assert set(targetids) == set(truth['TARGETID'])
 
     #- sort truth to match order of input targetids
+    # it doesn't matter if objtruth is sorted
     if len(targetids) == len(truth['TARGETID']):
         i = np.argsort(targetids)
         j = np.argsort(truth['TARGETID'])
@@ -858,7 +875,7 @@ def read_mock_spectra(truthfile, targetids, mockdir=None):
     wave = desispec.io.util.native_endian(wave).astype(np.float64)
     reordered_flux = desispec.io.util.native_endian(reordered_flux).astype(np.float64)
 
-    return reordered_flux, wave, reordered_truth
+    return reordered_flux, wave, reordered_truth, objtruth
 
 def targets2truthfiles(targets, basedir, nside=64):
     '''