+++ doc/changes.rst
@@ -5,6 +5,7 @@ desisim change log
 0.30.1 (unreleased)
 -------------------
 
+* Read and write `select_mock_targets` style `simspec` file #416 `PR #416`_). 
 * Add optional `nside` and `overwrite` arguments to `wrap-newexp` and
   `obs.new_exposure`, respectively (`PR #412`_).
 * Major (and backwards-incompatible) refactor of how the template/simulated
@@ -12,6 +13,7 @@ desisim change log
 
 .. _`PR #409`: https://github.com/desihub/desisim/pull/409
 .. _`PR #412`: https://github.com/desihub/desisim/pull/412
+.. _`PR #416`: https://github.com/desihub/desisim/pull/416
 
 0.30.0 (2018-08-09)
 -------------------